add nginx to the docker compose
Because of the auth server (and probably some connections to the file
server), we need to be able to route traffic from the main port to its
relevant server. I could have used HaProxy (and given I now understand
Docker networking more than basically nothing), but it was giving me
grief and I didn't like the syntax.