Disable verification on signup
Since this will be run as a private server, I don't need to worry about
actually verifying that people own FFXIV or anything. At most this code
will likely be ran once when I make the account initially, and every
other code will be a secondary that won't need verification.